Split app to main and public apps
